requirements.txt
numpy>=1.19
opencv_python>=4.8.1
opencv_python_headless==4.8.0.74
pandas>=1.1
Pillow>=8.4
torch>=1.10
torchvision>=0.11
facenet_pytorch>=2.5.3
matplotlib>=3.3.4
yolov5==6.0.3
scikit_learni>=0.20
